refactor: extract row sorting helper
